pyproject.toml
[project]
name = "news-picker"
version = "0.1.0"
description = "朝の情報読み上げシステム"
requires-python = ">=3.10"
dependencies = [
    "feedparser>=6.0.12",
    "google-genai>=1.65.0",
    "gtts>=2.5.4",
    "pygame>=2.6.1",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]
